backend/requirements.txt
# PATH: backend/requirements.txt
# PURPOSE: Python dependencies for FSE Accounting FastAPI backend

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AWS
boto3==1.34.25
botocore==1.34.25

# DocuSign
docusign-esign==3.25.0

# Stripe
stripe==7.10.0

# AI / DeepSeek
httpx==0.26.0
openai==1.10.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

